Split Mapper into two, more specialized classes
